+++ plugins/modules/fortiflexvm_tools_calc_info.py
@@ -438,9 +438,14 @@ options:
                 type: int
                 required: true
             dedicatedIPs:
-                description: Number between 4 and 65,534 (inclusive).
+                description: Number between 4 and 65,534 (inclusive). Value should be divisible by 4.
                 type: int
                 required: true
+            computeRegion:
+                description:
+                    - Additional Compute Region. Number between 0 and 16 (inclusive).
+                    - It can be scaled up in an increment of 1 but scaling down is NOT allowed.
+                type: int
     fortiEDR:
         description:
             - fortiEDR Cloud Configuration.
@@ -459,8 +464,52 @@ options:
                 type: list
                 elements: str
                 default: []
+    fortiRecon:
+        description:
+            - fortiRecon Cloud Configuration.
+        type: dict
+        suboptions:
+            service:
+                description:
+                    - Service package.
+                    - FRNEASM (External Attack Surface Monitoring);
+                    - FRNEASMBP (External Attack Surface Monitoring & Brand Protect);
+                    - FRNEASMBPACI (External Attack Surface Monitoring & Brand Protect & Adversary Centric Intelligence).
+                type: str
+                required: true
+            assets:
+                description: Number of Monitored Assets. Number between 200 and 1,000,000 (inclusive). Value should be divisible by 50.
+                type: int
+                required: true
+            networks:
+                description: Internal Attack Surface Monitoring (number of networks). Number between 0 and 100 (inclusive).
+                type: int
+            executives:
+                description: Executive Monitoring (number of executives). Number between 0 and 1,000 (inclusive).
+                type: int
+            vendors:
+                description: Vendor Monitoring (number of vendors). Number between 0 and 1,000 (inclusive).
+                type: int
+    fortiSIEMCloud:
+        description:
+            - fortiSIEM Cloud Configuration.
+        type: dict
+        suboptions:
+            computeUnits:
+                description: Number of Compute Units. Number between 10 and 600 (inclusive).
+                type: int
+                required: true
+            onlineStorage:
+                description:
+                    - Additional Online Storage. Number between 500 and 60,000 (inclusive). Value should be divisible by 500.
+                    - It can be scaled up in an increment of 500 but scaling down is NOT allowed.
+                type: int
+            archiveStorage:
+                description:
+                    - Archive Storage. Number between 0 and 60,000 (inclusive). Value should be divisible by 500.
+                    - can be scaled up in an increment of 500 but scaling down is NOT allowed.
+                type: int
 '''
-
 
 EXAMPLES = '''
 - name: Estimate cost
@@ -577,12 +626,28 @@ EXAMPLES = '''
         #   users: 50                         # 50 ~ 50000. Value should be divisible by 25.
         #   service: "FSASESTD"               # "FSASESTD" (Standard) or "FSASEADV" (Advanced).
         #   bandwidth: 100                    # 25 ~ 10000. Value should be divisible by 25.
-        #   dedicatedIPs: 10                  # 4 ~ 65534
+        #   dedicatedIPs: 12                  # 4 ~ 65534.  Value should be divisible by 4.
+        #   computeRegion: 1                  # 0 ~ 16. It can be scaled up in an increment of 1 but scaling down is NOT allowed.
 
         # fortiEDR:
         #   service: "FEDRPDR"                # "FEDRPDR" (Discover/Protect/Respond)
         #   endpoints: 10                     # Value should be between 0 and 50000
         #   addons: ["FEDRXDR"]               # Empty list or "FEDRXDR"
+
+        # fortiRecon:
+        #   service: "FRNEASM"                # "FRNEASM", "FRNEASMBP", "FRNEASMBPACI"
+        #   assets: 200                       # 200 ~ 1000000. Value should be divisible by 50
+        #   networks: 0                       # 0 ~ 100
+        #   executives: 0                     # 0 ~ 1000
+        #   vendors: 0                        # 0 ~ 100
+
+        # fortiSIEMCloud:
+        #   computeUnits: 10                  # 10 ~ 600
+        #   onlineStorage: 500                # 500 ~ 60000. Value should be divisible by 500.
+        #                                     # It can be scaled up in an increment of 500 but scaling down is NOT allowed.
+        #   archiveStorage: 0                 # 0 ~ 60000. Value should be divisible by 500.
+        #                                     # It can be scaled up in an increment of 500 but scaling down is NOT allowed.
+
       register: result
 
     - name: Display response